requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
redis==5.0.1
pika==1.3.2
python-dotenv==1.0.0
httpx==0.25.1
tenacity==8.2.3
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
aio-pika==9.4.1